make spectra a standalone module
