+++ SAGA/database/__init__.py
@@ -7,4 +7,3 @@ This subpackage contains database-related routines, including Database
 from .core import *
 from .saga_database import *
 from .external import *
-from .spectra import SpectraData
