+++ SAGA/version.py
@@ -1,4 +1,4 @@
 """
 SAGA package version
 """
-__version__ = '0.7.0'
+__version__ = '0.8.0'
